Preparationbookingcommunication

+++ Presentation/Models/Invoice.cs
@@ -14,7 +14,7 @@ public class Invoice
 
     public DateTime DueDate { get; set; }
 
-    public int TotalAmount { get; set; }
+    public decimal TotalAmount { get; set; }
 
     public string Status { get; set; } = null!;
